refactor(envelope): split finding owner
